feat: User Joi 유효성 추가, 그룹 및 클라이언트 조회 정렬 변경, Swagger Set up

+++ src/services/group.service.js
@@ -1,5 +1,9 @@
 const { logger } = require('../middlewares/logger');
-const { BadRequestError, NotFoundError } = require('../exceptions/errors');
+const {
+  BadRequestError,
+  NotFoundError,
+  ForbiddenError,
+} = require('../exceptions/errors');
 const GroupRepository = require('../repositories/group.repository');
 
 module.exports = class GroupService {
@@ -7,19 +11,15 @@ module.exports = class GroupService {
     this.groupRepository = new GroupRepository();
   }
   // 빈 Group 생성
-  createGroup = async ({
-    //userId,
-    groupName,
-    groupDescription,
-  }) => {
+  createGroup = async ({ userId, companyId, groupName, groupDescription }) => {
     logger.info(`GroupService.createGroup Request`);
 
     const groupData = await this.groupRepository.createGroup({
-      // userId,
+      userId,
+      companyId,
       groupName,
       groupDescription,
     });
-
     if (!groupData) {
       throw new BadRequestError('그룹 생성에 실패하였습니다.');
     }
@@ -29,6 +29,7 @@ module.exports = class GroupService {
   //그룹 전체 조회
   getAllGroup = async () => {
     logger.info(`GroupService.getAllGroup Request`);
+
     const allGroupData = await this.groupRepository.getAllGroup();
     if (!allGroupData) {
       throw new BadRequestError('그룹 조회에 실패하였습니다.');
@@ -37,13 +38,22 @@ module.exports = class GroupService {
   };
 
   //그룹 삭제
-  deleteGroup = async ({ groupId }) => {
+  deleteGroup = async ({ userId, companyId, groupId }) => {
     logger.info(`GroupService.deleteGroup Request`);
-    const findGroupData = await this.groupRepository.findGroupId({ groupId });
+    const findGroupData = await this.groupRepository.findGroupId({
+      userId,
+      companyId,
+      groupId,
+    });
     if (!findGroupData) {
       throw new NotFoundError('그룹 삭제에 실패하였습니다.');
     }
+    if (findGroupData.userId !== userId) {
+      throw new ForbiddenError('접근 권한이 없습니다.');
+    }
     const deleteGroupData = await this.groupRepository.deleteGroup({
+      userId,
+      companyId,
       groupId,
     });
